requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.3.1
    # via kombu
asgiref==3.9.1
    # via django
billiard==4.2.1
    # via celery
celery==5.5.3
    # via
    #   -r requirements.in
    #   django-celery-beat
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
cron-descriptor==1.4.5
    # via django-celery-beat
django==5.2.4
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
django-celery-beat==2.8.1
    # via -r requirements.in
django-timezone-field==7.1
    # via django-celery-beat
djangorestframework==3.16.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
idna==3.10
    # via requests
kombu==5.5.4
    # via celery
mysqlclient==2.2.7
    # via -r requirements.in
packaging==25.0
    # via
    #   gunicorn
    #   kombu
prompt-toolkit==3.0.51
    # via click-repl
pyjwt==2.10.1
    # via djangorestframework-simplejwt
python-crontab==3.3.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.1.1
    # via -r requirements.in
redis==6.2.0
    # via -r requirements.in
requests==2.32.4
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
tzdata==2025.2
    # via
    #   django-celery-beat
    #   kombu
urllib3==2.5.0
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
whitenoise==6.9.0
    # via -r requirements.in
